+++ src/trx/game/ui/dialogs/photo_mode.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <trx/game/photo_mode.h>
-#include <trx/game/ui/common.h>
-
-// A photo mode tutorial dialog.
-
-void UI_PhotoMode(PHOTO_MODE current_mode);